orders订单表减少用户联系方式phonenumber和用户地址address1时，增加Users类及相应查询语句，在users表里面查询phonenumber和address1，将其返回到userList界面
