+++ src/main/java/com/example/controller/DetailController.java
@@ -2,8 +2,10 @@ package com.example.controller;
 
 import com.example.entity.Detail;
 import com.example.entity.Order;
+import com.example.entity.Users;
 import com.example.service.DetailService;
 import com.example.service.OrderService;
+import com.example.service.UsersService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -16,9 +18,15 @@ public class DetailController {
     private DetailService detailService;
     @Autowired
     private OrderService orderService;
+    @Autowired
+    private UsersService usersService;
 
     @RequestMapping("/topaysucessful")
-    public String topaysucessful(Model model,Order order,Detail detail){
+    public String topaysucessful(Model model, Order order, Detail detail, Users users){
+        users = usersService.findusers();
+        model.addAttribute("users",users);
+
+
         order = orderService.findOrder1();
         model.addAttribute("order",order);
         detail.setUsername(order.getUsername());
