+++ src/main/java/com/example/dao/UsersMapper.java
@@ -1,0 +1,7 @@
+package com.example.dao;
+
+import com.example.entity.Users;
+
+public interface UsersMapper {
+    Users findusers();
+}
