+++ src/main/java/com/example/entity/Users.java
@@ -1,0 +1,22 @@
+package com.example.entity;
+
+public class Users {
+    private String phonenumber;
+    private String address1;
+
+    public String getPhonenumber() {
+        return phonenumber;
+    }
+
+    public void setPhonenumber(String phonenumber) {
+        this.phonenumber = phonenumber;
+    }
+
+    public String getAddress1() {
+        return address1;
+    }
+
+    public void setAddress1(String address1) {
+        this.address1 = address1;
+    }
+}
