+++ src/main/java/com/example/service/UsersService.java
@@ -1,0 +1,7 @@
+package com.example.service;
+
+import com.example.entity.Users;
+
+public interface UsersService {
+    Users findusers();
+}
