+++ src/main/resources/templates/paysucessful.html
@@ -141,10 +141,10 @@
             <div class="box2">
                 <ul class="outer">
 
-                    <li> <tr th:each="order : ${order}"><!--循环语句，条件为users列表长度-->
-                        收货地址：<p3 th:text="${order.address1}"></p3><br>
+                    <li> <!--循环语句，条件为users列表长度-->
+                        收货地址：<p3 th:text="${users.address1}"></p3><br>
                         <p3 th:text="${order.username}"></p3>
-                        <p3 th:text="${order.phonenumber}"></p3><br>
+                        <p3 th:text="${users.phonenumber}"></p3><br>
                         实付款：￥<p3 th:text="${order.gPrice*order.orderNum}"></p3></tr>
                     </li>
                     <li></li>
